Allow decks to be marked as public. Make Deck color input select.
Reintroduce the "mark deck as public" button that was lost in the
previous refactor.

In the deck creation modal, the input for the color attribute is now a
select.

Changed the Form reusable component to use the dot syntax for child
compoenents e.g. Form.Input and Form.Select

Introduce the Form.Select child component of the Form reusable component